move express config to the presentation layer
